+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1
+
+- Changed namespace from `Jens\ImageConverter` to `ImageConverter`
+- Added `convert` helper function
+- Changed `example.php` to include the shorter version
+
 ## 1.0
 
 - Initial release